[BB-1244] Fix issue with circular imports
